pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "vton3d-bat"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "numpy==1.26.1",
    "Pillow",
    "huggingface_hub",
    "transformers",
    "einops",
    "safetensors",
    "jupyter",
    "gradio==5.17.1",
    "viser==0.2.23",
    "tqdm",
    "hydra-core",
    "omegaconf",
    "opencv-python",
    "scipy",
    "onnxruntime",
    "requests",
    "accelerate>=0.26.0",
    "bitsandbytes",
    "trimesh",
    "matplotlib",
    "torchmetrics[image]",
    "pydantic==2.10.6",
    "pycolmap==3.10.0",
    "pyceres==2.3",

    # Git dependencies
    "lightglue @ git+https://github.com/jytime/LightGlue.git",
    "vggt @ git+https://github.com/vtonSD-bat/vggt.git",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["vton3d",]
